+++ api/account_opening.go
@@ -12,11 +12,11 @@ func NewAccountOpeningService(c *Client) *AccountOpeningService {
 	return &AccountOpeningService{client: c}
 }
 
-func (svc *AccountOpeningService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountOpeningService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountOpeningModel, name)
 }
 
-func (svc *AccountOpeningService) GetByIds(ids []int) (*types.AccountOpenings, error) {
+func (svc *AccountOpeningService) GetByIds(ids []int64) (*types.AccountOpenings, error) {
 	a := &types.AccountOpenings{}
 	return a, svc.client.getByIds(types.AccountOpeningModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountOpeningService) GetAll() (*types.AccountOpenings, error) {
 	return a, svc.client.getAll(types.AccountOpeningModel, a)
 }
 
-func (svc *AccountOpeningService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountOpeningService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountOpeningModel, fields, relations)
 }
 
-func (svc *AccountOpeningService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountOpeningService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountOpeningModel, ids, fields, relations)
 }
 
-func (svc *AccountOpeningService) Delete(ids []int) error {
+func (svc *AccountOpeningService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountOpeningModel, ids)
 }
