+++ api/mail_activity.go
@@ -12,11 +12,11 @@ func NewMailActivityService(c *Client) *MailActivityService {
 	return &MailActivityService{client: c}
 }
 
-func (svc *MailActivityService) GetIdsByName(name string) ([]int, error) {
+func (svc *MailActivityService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.MailActivityModel, name)
 }
 
-func (svc *MailActivityService) GetByIds(ids []int) (*types.MailActivitys, error) {
+func (svc *MailActivityService) GetByIds(ids []int64) (*types.MailActivitys, error) {
 	m := &types.MailActivitys{}
 	return m, svc.client.getByIds(types.MailActivityModel, ids, m)
 }
@@ -36,14 +36,14 @@ func (svc *MailActivityService) GetAll() (*types.MailActivitys, error) {
 	return m, svc.client.getAll(types.MailActivityModel, m)
 }
 
-func (svc *MailActivityService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *MailActivityService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.MailActivityModel, fields, relations)
 }
 
-func (svc *MailActivityService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *MailActivityService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.MailActivityModel, ids, fields, relations)
 }
 
-func (svc *MailActivityService) Delete(ids []int) error {
+func (svc *MailActivityService) Delete(ids []int64) error {
 	return svc.client.delete(types.MailActivityModel, ids)
 }
