+++ api/mail_activity_mixin.go
@@ -12,11 +12,11 @@ func NewMailActivityMixinService(c *Client) *MailActivityMixinService {
 	return &MailActivityMixinService{client: c}
 }
 
-func (svc *MailActivityMixinService) GetIdsByName(name string) ([]int, error) {
+func (svc *MailActivityMixinService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.MailActivityMixinModel, name)
 }
 
-func (svc *MailActivityMixinService) GetByIds(ids []int) (*types.MailActivityMixins, error) {
+func (svc *MailActivityMixinService) GetByIds(ids []int64) (*types.MailActivityMixins, error) {
 	m := &types.MailActivityMixins{}
 	return m, svc.client.getByIds(types.MailActivityMixinModel, ids, m)
 }
@@ -36,14 +36,14 @@ func (svc *MailActivityMixinService) GetAll() (*types.MailActivityMixins, error)
 	return m, svc.client.getAll(types.MailActivityMixinModel, m)
 }
 
-func (svc *MailActivityMixinService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *MailActivityMixinService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.MailActivityMixinModel, fields, relations)
 }
 
-func (svc *MailActivityMixinService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *MailActivityMixinService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.MailActivityMixinModel, ids, fields, relations)
 }
 
-func (svc *MailActivityMixinService) Delete(ids []int) error {
+func (svc *MailActivityMixinService) Delete(ids []int64) error {
 	return svc.client.delete(types.MailActivityMixinModel, ids)
 }
