+++ api/project_task_merge_wizard.go
@@ -12,11 +12,11 @@ func NewProjectTaskMergeWizardService(c *Client) *ProjectTaskMergeWizardService 
 	return &ProjectTaskMergeWizardService{client: c}
 }
 
-func (svc *ProjectTaskMergeWizardService) GetIdsByName(name string) ([]int, error) {
+func (svc *ProjectTaskMergeWizardService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.ProjectTaskMergeWizardModel, name)
 }
 
-func (svc *ProjectTaskMergeWizardService) GetByIds(ids []int) (*types.ProjectTaskMergeWizards, error) {
+func (svc *ProjectTaskMergeWizardService) GetByIds(ids []int64) (*types.ProjectTaskMergeWizards, error) {
 	p := &types.ProjectTaskMergeWizards{}
 	return p, svc.client.getByIds(types.ProjectTaskMergeWizardModel, ids, p)
 }
@@ -36,14 +36,14 @@ func (svc *ProjectTaskMergeWizardService) GetAll() (*types.ProjectTaskMergeWizar
 	return p, svc.client.getAll(types.ProjectTaskMergeWizardModel, p)
 }
 
-func (svc *ProjectTaskMergeWizardService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *ProjectTaskMergeWizardService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.ProjectTaskMergeWizardModel, fields, relations)
 }
 
-func (svc *ProjectTaskMergeWizardService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *ProjectTaskMergeWizardService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.ProjectTaskMergeWizardModel, ids, fields, relations)
 }
 
-func (svc *ProjectTaskMergeWizardService) Delete(ids []int) error {
+func (svc *ProjectTaskMergeWizardService) Delete(ids []int64) error {
 	return svc.client.delete(types.ProjectTaskMergeWizardModel, ids)
 }
