+++ api/report_account_report_tax.go
@@ -12,11 +12,11 @@ func NewReportAccountReportTaxService(c *Client) *ReportAccountReportTaxService 
 	return &ReportAccountReportTaxService{client: c}
 }
 
-func (svc *ReportAccountReportTaxService) GetIdsByName(name string) ([]int, error) {
+func (svc *ReportAccountReportTaxService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.ReportAccountReportTaxModel, name)
 }
 
-func (svc *ReportAccountReportTaxService) GetByIds(ids []int) (*types.ReportAccountReportTaxs, error) {
+func (svc *ReportAccountReportTaxService) GetByIds(ids []int64) (*types.ReportAccountReportTaxs, error) {
 	r := &types.ReportAccountReportTaxs{}
 	return r, svc.client.getByIds(types.ReportAccountReportTaxModel, ids, r)
 }
@@ -36,14 +36,14 @@ func (svc *ReportAccountReportTaxService) GetAll() (*types.ReportAccountReportTa
 	return r, svc.client.getAll(types.ReportAccountReportTaxModel, r)
 }
 
-func (svc *ReportAccountReportTaxService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *ReportAccountReportTaxService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.ReportAccountReportTaxModel, fields, relations)
 }
 
-func (svc *ReportAccountReportTaxService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *ReportAccountReportTaxService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.ReportAccountReportTaxModel, ids, fields, relations)
 }
 
-func (svc *ReportAccountReportTaxService) Delete(ids []int) error {
+func (svc *ReportAccountReportTaxService) Delete(ids []int64) error {
 	return svc.client.delete(types.ReportAccountReportTaxModel, ids)
 }
