+++ api/stock_warn_insufficient_qty_scrap.go
@@ -12,11 +12,11 @@ func NewStockWarnInsufficientQtyScrapService(c *Client) *StockWarnInsufficientQt
 	return &StockWarnInsufficientQtyScrapService{client: c}
 }
 
-func (svc *StockWarnInsufficientQtyScrapService) GetIdsByName(name string) ([]int, error) {
+func (svc *StockWarnInsufficientQtyScrapService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.StockWarnInsufficientQtyScrapModel, name)
 }
 
-func (svc *StockWarnInsufficientQtyScrapService) GetByIds(ids []int) (*types.StockWarnInsufficientQtyScraps, error) {
+func (svc *StockWarnInsufficientQtyScrapService) GetByIds(ids []int64) (*types.StockWarnInsufficientQtyScraps, error) {
 	s := &types.StockWarnInsufficientQtyScraps{}
 	return s, svc.client.getByIds(types.StockWarnInsufficientQtyScrapModel, ids, s)
 }
@@ -36,14 +36,14 @@ func (svc *StockWarnInsufficientQtyScrapService) GetAll() (*types.StockWarnInsuf
 	return s, svc.client.getAll(types.StockWarnInsufficientQtyScrapModel, s)
 }
 
-func (svc *StockWarnInsufficientQtyScrapService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *StockWarnInsufficientQtyScrapService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.StockWarnInsufficientQtyScrapModel, fields, relations)
 }
 
-func (svc *StockWarnInsufficientQtyScrapService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *StockWarnInsufficientQtyScrapService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.StockWarnInsufficientQtyScrapModel, ids, fields, relations)
 }
 
-func (svc *StockWarnInsufficientQtyScrapService) Delete(ids []int) error {
+func (svc *StockWarnInsufficientQtyScrapService) Delete(ids []int64) error {
 	return svc.client.delete(types.StockWarnInsufficientQtyScrapModel, ids)
 }
